moved data processing to own files
