+++ src/App.js
@@ -4,59 +4,10 @@ import { csv } from 'd3'
 import { Container, Grid } from '@material-ui/core';
 import { DoubleLineChart } from './components/DoubleLineChart.js'
 import { FilterSelect } from './components/FilterSelect.js'
-import { stringToDate } from './utilities/stringToDate.js'
-import { incrementDateString } from './utilities/incrementDateString.js'
+import { createChartData } from './data/createChartData.js'
+import { uniquePropertyValuesSorted } from './data/uniquePropertyValuesSorted.js'
+import { filterByPropertyValue } from './data/filterByPropertyValue.js'
 import './App.css';
-
-// App ------------------------------------------
-
-const filterByPropertyValue = (data, key, values) => {
-  return values.length ? data.filter(d => _.includes(values, d[key])) : data;
-}
-
-const uniquePropertyValuesSorted = (data, key) => {
-  return _.uniq(data.map(d => d[key])).sort()
-}
-
-const accumulatedNumberOrNaN = (list, key) => {
-  if (_.findIndex(list, e => e[key] ) !== -1)
-  {
-    return _.sumBy(list, e => { return e[key].length ? Number.parseInt(e[key]) : 0 })
-  }
-  return NaN
-}
-
-const createDatesListRecursive = (datesList, currentDate, lastDate) => {
-  datesList.push({ name: currentDate });
-  if (currentDate !== lastDate)
-  {
-    createDatesListRecursive(datesList, incrementDateString(currentDate), lastDate)
-  }
-  return datesList;
-}
-
-const createChartData = (data, currentDataSources, currentCampaigns) => {
-  const filteredData = filterByPropertyValue(
-    filterByPropertyValue(
-      data,
-      "Datasource",
-      currentDataSources
-    ),
-    "Campaign",
-    currentCampaigns
-  );
-
-  const dates = _.map(_.groupBy(filteredData, "Date"), d => {
-    return {
-      name: _.first(d).Date,
-      Clicks : accumulatedNumberOrNaN(d, "Clicks"),
-      Impressions : accumulatedNumberOrNaN(d, "Impressions")
-    }
-  })
-
-  const allDates = data.length ? createDatesListRecursive([], _.first(data).Date, _.last(data).Date) : {}
-  return _.sortBy(_.unionBy(dates, allDates, "name"), [d => stringToDate(d.name)]);
-}
 
 const App = () => {
   const [data, setData] = React.useState([])
@@ -71,8 +22,7 @@ const App = () => {
     []
   )
 
-  const chartData = React.useMemo(
-    () => {
+  const chartData = React.useMemo(() => {
       return createChartData(data, currentDataSources, currentCampaigns)
     },
     [data, currentDataSources, currentCampaigns]
@@ -84,8 +34,7 @@ const App = () => {
     [data, currentCampaigns]
   )
 
-  const availableCampaigns = React.useMemo(
-    () => {
+  const availableCampaigns = React.useMemo(() => {
       return uniquePropertyValuesSorted(filterByPropertyValue(data, "Datasource", currentDataSources), "Campaign")
     },
     [data, currentDataSources]
